Alterações no front para a modificação de prefeitura funcionar

+++ src/main/java/com/ufcg/si1/service/interfaces/QueixaService.java
@@ -21,6 +21,6 @@ public interface QueixaService {
 
 	public void deletaTodasQueixas();
 	
-	public long quantidadeQueixasAbertas();
+	public double quantidadeQueixasAbertas();
 
 }
